AS, RX200, fixed a casting error in Sci_Internal.h
git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@2710 88d5ba13-baa4-4915-ae7a-41304cef7dd7

+++ contrib/as/arch/rx200_5f5210x/ccrx/include/Sci_Internal.h
@@ -92,7 +92,7 @@
 	(_ch != SCI_CHANNEL_12 ? HW_SCI_C_BASE_ADDR : HW_SCI_D_BASE_ADDR  )
 
 /* Channel Hardware Registers Read/Write */
-#define	UART_REG(_ch, _ofs)	EE_HWREG8( (uint32) (UART_GET_BASE_ADDR(_ch)) | \
+#define	UART_REG(_ch, _ofs)	EE_HWREG8( (uint8) (UART_GET_BASE_ADDR(_ch)) | \
 		(_ch * HW_SCI_CH_OFFSET) | _ofs)
 
 #define	UART_REG_SET(_ch, _ofs, _val) (UART_REG(_ch,_ofs) = _val)
